+++ CareerLink/urls.py
@@ -18,7 +18,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('accounts.urls')),
-    path('accounts/', include('accounts.urls')),
+    path('', include('jobs.urls')),
     path('jobs/', include('jobs.urls'))
 ]
