--- accounts/admin.py
@@ -1,28 +1,0 @@
-from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin
-
-from .forms import CustomUserCreationForm, CustomUserChangeForm
-from .models import CustomUser
-
-
-class CustomUserAdmin(UserAdmin):
-    add_form = CustomUserCreationForm
-    form = CustomUserChangeForm
-    model = CustomUser
-    list_display = ('email', 'first_name', 'last_name', 'is_staff', 'is_active',)
-    list_filter = ('email','first_name', 'last_name', 'is_staff', 'is_active',)
-    fieldsets = (
-        (None, {'fields': ('email', 'first_name', 'last_name', 'password')}),
-        ('Permissions', {'fields': ('is_staff', 'is_active')}),
-    )
-    add_fieldsets = (
-        (None, {
-            'classes': ('wide',),
-            'fields': ('email', 'first_name', 'last_name', 'password1', 'password2', 'is_staff', 'is_active')}
-        ),
-    )
-    search_fields = ('email',)
-    ordering = ('email',)
-
-
-admin.site.register(CustomUser, CustomUserAdmin)
