--- accounts/templates/index.html
@@ -1,24 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<div class="container" style="margin-top:5%; margin-left: 30%">
-  <div class="card" style="width:600px; height: 400px; ">
-    <div class="container text-white bg-info" style = "text-align: center">
-      <h1>Welcome</h1>
-    </div>
-    <div class="card-body" style="margin-top: 0px;margin-left: 0px;margin-bottom: 0px;border-left: 100px">
-      <p class="card" style="margin-left: 0px; margin-top: 0px;">
-        <div class="btn-group-vertical">
-          {% if user.is_authenticated %}
-          <a class="btn btn-primary" style="width: 120px;margin-top:30%;margin-left:220px"; href="{% url 'dashboard_url' %}">Dashboard</a>
-          <a class="btn btn-primary" style="width: 120px;margin-top:1px;margin-left: 220px" href="{% url 'logout_url' %}">Logout</a>
-          {% else %}
-          <a class="btn btn-primary" style="width: 100px;margin-top:30%;margin-left: 220px" href="{% url 'login_url' %}">Login</a>
-          <br>
-          <a class="btn btn-primary" style="width: 100px;margin-bottom:15%;margin-left: 220px"; href="{% url 'register_url' %}">Register</a>
-          {% endif %}
-        </div>
-      </p>
-    </div>
-  </div>
-</div>
-{% endblock %}
