--- accounts/templates/search.html
@@ -1,25 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-
-
-
-
-<div class="container" style="margin-top:5%; margin-left: 30%">
-    <div class="card" style="width:600px; height: 400px; ">
-        <div class="container text-white bg-danger" style = "text-align: center">
-            <h1>Find Job Near You</h1>
-        </div>
-        <div class="card-body" style="margin-top: 0px;margin-left: 0px;margin-bottom: 0px;border-left: 100px">
-            <p class="card" style="margin-left: 0px; margin-top: 0px;">
-            <div class=“w-100” style="margin: 0 20% 0 20% ">
-                <input class="form-control mr-sm-2" type="search" placeholder="Job" aria-label="Job" style="margin:20% 50% 0 0% ">
-                <input class="form-control mr-sm-2" type="search" placeholder="Location" aria-label="Location" style="margin: 10% 20% 10% 0 ">
-                <button class="btn btn-outline-warning my-2 my-sm-0" type="submit" bg-dark >Search</button>
-            </div>
-            </p>
-        </div>
-    </div>
-</div>
-
-{% endblock %}
