+++ jobs/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
 
 
-class JobsConfig(AppConfig):
-    name = 'jobs'
+class AccountsConfig(AppConfig):
+    name = 'accounts'
