+++ jobs/managers.py
