+++ jobs/templates/dashboard.html
@@ -1,20 +1,17 @@
 {% extends "base.html" %}
 {% block content %}
-  <body>
-      <div class="alert alert-success" role="alert" style="margin-top: 10%;margin-left: 20%;margin-right: 15%">
-          <h4 class="alert-heading">Welcome, {{user.first_name}}</h4>
-          <p>Aww yeah, you successfully read this registration successful message</p>
-          <hr>
-          <div class="container" style="margin-left: 30%">
-              <div class="pik-body">
-                  <img class="shock-left" src="http://www.megghy.com/gif_animate/natura/tuoni/3.gif">
-                  <img class="shock-right" src="http://www.megghy.com/gif_animate/natura/tuoni/3.gif">
-
-              </div>
-          </div>
-          <hr>
-          <button class="btn btn-primary" style="width: 120px;margin-top:0%;margin-left:40%" href="{% url 'logout_url' %}">Logout</button>
-      </div>
-  </body>
+    <hr>
+    <div style="float: left;width: 45%">
+        {% for post in posts %}
+            <h1>{{ post.title }}</h1>
+            <p>By {{ post.author.first_name }} {{ post.author.last_name }} on {{ post.date }}</p>
+            <p>{{ post.requirements }}</p>
+            <p>{{ post.description }}</p>
+            <p>Contact: {{ post.author.email }}</p>
+            <hr>
+        {% endfor %}
+    </div>
+    <div style="border: solid; margin-left: 50%;height: 100%">
+        <h3> Post New Job Form</h3>
+    </div>
 {% endblock %}
-
