+++ jobs/templates/index.html
@@ -1,0 +1,17 @@
+{% extends "base.html" %}
+
+{% block content %}
+  {% if user.is_authenticated %}
+    <a href="{% url 'dashboard_url' %}">Dashboard</a>
+    <a href="{% url 'logout_url' %}">Logout</a>
+  {% else %}
+    <div class = "card" style="margin: 5% 30%;">
+    <div class = "card-title text-white bg-success" style="text-align: center;background-color: purple-gradient">
+    <h1>Welcome</h1>
+    </div>
+    <div style="margin: 10%;text-align: center">
+    <a href="{% url 'login_url' %}">Login</a>
+    <a href="{% url 'register_url' %}">Register</a>
+    </div>
+  {% endif %}
+{% endblock %}
